Restructured the components folder. Added a nav and auth folder to hold respective components. In nav folder also added a buttons folder to hold different nav-bar button components
