+++ client/src/App.js
@@ -3,9 +3,9 @@ import { BrowserRouter, Switch, Route, useLocation } from 'react-router-dom';
 import {useSelector} from 'react-redux';
 
 import UserList from './components/UsersList';
-import Login from './components/Login';
-import NavBar from './components/NavBar';
-import Signup from './components/Signup';
+import Login from './components/auth/Login';
+import NavBar from './components/nav/NavBar';
+import Signup from './components/auth/Signup';
 
 function App() {
     //Will be either undefined or a number depending on if there is a token in the cookies
@@ -25,10 +25,6 @@ function App() {
 
                 <Route path="/signup">
                     <Signup />
-                </Route>
-
-                <Route path="/users">
-                    <UserList />
                 </Route>
 
                 <Route path="/">
