+++ client/src/index.css
@@ -43,6 +43,15 @@ html, body, #root, .App {
   align-items: center;
 }
 
+.nav-buttons {
+  display: flex;
+  margin: auto;
+}
+
+.auth-nav-buttons {
+  margin: auto;
+}
+
 .logo {
   color: brown;
   font-size: 20px;
